backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email_validator
